LocalApp/requirements.txt
Flask
cryptography
python-dotenv
# Note: pqcrypto không khả dụng trên Windows, đã sử dụng mock implementation
# pqcrypto 